+++ tests/UiTests/NurseryTests/04_PreNurseryCulling.spec.js
@@ -48,11 +48,8 @@ test.describe.serial("Pre Nursery Culling Tests", () => {
 
     await checkLength(paths, columns, createValues, editValues);
 
-    // Clean up existing record if any
     docNo = DocNo[keyName];
-    if (docNo) {
-      await db.deleteData(deleteSQL, { DocNo: docNo });
-    }
+    if (docNo) await db.deleteData(deleteSQL, { DocNo: docNo });
 
     console.log(`Start Running: ${formName}`);
   });
@@ -135,5 +132,9 @@ test.describe.serial("Pre Nursery Culling Tests", () => {
   });
 
   // ---------------- After All ----------------
-  test.afterAll(async () => {});
+  test.afterAll(async ({ db }) => {
+    if (docNo) await db.deleteData(deleteSQL, { DocNo: docNo });
+
+    console.log(`End Running: ${formName}`);
+  });
 });
